backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "django-filter>=24.2",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "environs[django]>=14.1.1",
    "gdal>=3.6.0,<3.7.0",
    "gunicorn>=23.0.0",
    "hmmlearn>=0.3.3",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "coverage==7.4.0",
    "factory-boy==3.3.0",
    "faker==22.0.0",
    "model-bakery==1.17.0",
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "pytest-django==4.8.0",
]
